调整用户管理模块逻辑；新增角色管理

+++ src/main/java/org/alex/platform/service/impl/UserServiceImpl.java
@@ -1,15 +1,20 @@
 package org.alex.platform.service.impl;
 
+import org.alex.platform.common.Result;
 import org.alex.platform.exception.BusinessException;
 import org.alex.platform.mapper.UserMapper;
 import org.alex.platform.pojo.UserDO;
+import org.alex.platform.pojo.UserVO;
 import org.alex.platform.service.UserService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import org.alex.platform.util.ValidUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Date;
 
 
 @Service
@@ -26,7 +31,7 @@ public class UserServiceImpl implements UserService {
      * @return PageInfo<UserDO>
      */
     @Override
-    public PageInfo<UserDO> findUserList(UserDO userDO, Integer pageNum, Integer pageSize) {
+    public PageInfo<UserVO> findUserList(UserDO userDO, Integer pageNum, Integer pageSize) {
         PageHelper.startPage(pageNum, pageSize);
         return new PageInfo<>(userMapper.selectUserList(userDO));
     }
@@ -41,6 +46,11 @@ public class UserServiceImpl implements UserService {
         if (!userMapper.checkUser(userDO).isEmpty()) {
             throw new BusinessException("用户名已存在");
         }
+        Integer userId = userDO.getUserId();
+        if (userId == 1) {
+            throw new BusinessException("内置用户禁止修改");
+        }
+        userDO.setUpdateTime(new Date());
         userMapper.updateUser(userDO);
     }
 
@@ -50,13 +60,48 @@ public class UserServiceImpl implements UserService {
      * @return
      */
     @Override
-    public Boolean saveUser(UserDO userDO) {
+    public void saveUser(UserDO userDO) throws BusinessException {
         String username = userDO.getUsername();
+        String password = userDO.getPassword();
+        Integer roleId = userDO.getRoleId();
+        Byte sex = userDO.getSex();
+        Date date = new Date();
+        userDO.setCreatedTime(date);
+        userDO.setUpdateTime(date);
+        userDO.setIsEnable((byte) 1);
+        if (sex == null || roleId == null || password == null) {
+            throw new BusinessException("请完善用户信息");
+        } else {
+            ValidUtil.length(password, 3, 20, "密码长度必须为3~20");
+            if (userMapper.selectUserByName(username) != null) {
+                throw new BusinessException("用户名已存在");
+            } else {
+                userMapper.insertUser(userDO);
+            }
+        }
+    }
+
+    /**
+     * 用户注册
+     * @param userDO 对象信息
+     * @throws BusinessException
+     */
+    @Override
+    public void registerUser(UserDO userDO) throws BusinessException {
+        String username = userDO.getUsername();
+        String password = userDO.getPassword();
+        Date date = new Date();
+        userDO.setCreatedTime(date);
+        userDO.setUpdateTime(date);
+        userDO.setIsEnable((byte) 1);
+        userDO.setRoleId(null);
+        userDO.setSex((byte) 1);
+        ValidUtil.notNUll(password, "请完善用户信息");
+        ValidUtil.length(password, 3, 20, "密码长度必须为3~20");
         if (userMapper.selectUserByName(username) != null) {
-            return false;
+            throw new BusinessException("用户名已存在");
         } else {
             userMapper.insertUser(userDO);
-            return true;
         }
     }
 
@@ -76,7 +121,7 @@ public class UserServiceImpl implements UserService {
      * @return UserDO
      */
     @Override
-    public UserDO findUserById(Integer userId) {
+    public UserVO findUserById(Integer userId) {
         return userMapper.selectUserById(userId);
     }
 
@@ -85,7 +130,10 @@ public class UserServiceImpl implements UserService {
      * @param userId 用户编号
      */
     @Override
-    public void removeUserById(Integer userId) {
+    public void removeUserById(Integer userId) throws BusinessException {
+        if (userId == 1) {
+            throw new BusinessException("内置用户禁止删除");
+        }
         userMapper.deleteUser(userId);
     }
 }
